% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_hola_nombre_edad.R
\name{func_hola_nombre_edad}
\alias{func_hola_nombre_edad}
\title{Esta funcion devuelve "Hola" + nombre + edad}
\usage{
func_hola_nombre_edad(nombre = "Mundo", edad = 27)
}
\description{
Esta funcion devuelve "Hola" + nombre + edad
}
