\name{RProgramming-package}
\alias{RProgramming-package}
\alias{RProgramming}
\docType{package}
\title{
\packageTitle{RProgramming}
}
\description{
\packageDescription{RProgramming}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RProgramming}
\packageIndices{RProgramming}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RProgramming}

Maintainer: \packageMaintainer{RProgramming}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
